put validation fun in module and apply it in app
